[BE-4679] Remove the collation trait from the list of trait definitions (#5220)
The default list of traits included in the planner is the collation
trait and the convention trait. We have now integrated the convention
trait and actively use it, but we haven't done the same with the
collation trait.

The collation trait is important for optimizations that involve the sort
order of the data. We don't presently have any of these. At the same
time, there's the possibility that plans could be created that rely on
this trait definition. Since we didn't integrate the collation trait
with the pandas versions of the operations, we may incorrectly not
preserve the trait and we might not insert the collation trait when this
is otherwise not possible.

The code for this is substantial enough that it's not necessarily
straightforward. Projection and filter were easy enough. Projection was
included in a previous commit and filter is included here, but there's
also joins and other more complicated operators that I can't be
confident will be correct.

Since it's not used and I'm not confident in the quality of the results
with the trait definition present, this removes it so we no longer
require it as part of the output and the VolcanoPlanner will not require
its presence or consider it when creating plans.

This was manually tested by reverting the projection change, verifying
it failed, then overriding the trait definitions and making sure it
passed without further changes.