Consolidate the test file location for java tests (#5235)
I originally attempted to have two different test source roots. When I
was attempting to get this working, I used one of the parent directories
as a root because otherwise JUnit wouldn't pick up both the java and
kotlin directory. At the same time, this attribute seemed to only allow
one directory instead of multiple directory roots.

This didn't really seem to work properly. Now, IntelliJ thinks that the
files are put into the wrong package.

Instead of overcomplicating this, just put everything into the Java
directory. The Kotlin compiler generally works fine if the files are
mixed with Java in the Java directory so this doesn't cause a problem.