[BSE-165] Implement computeSelfCost for PandasProject (#5259)
This implements `computeSelfCost` for `PandasProject` in a simple manner
to estimate how long that operation might take.

Specifically, references to `InputRef` are considered free because they
are essentially renames. Operations, which are usually represented by
calls, have a static small cost associated with them and a memory cost
associated with the creation of the new vector.

Some operations, like `OVER()` nodes, are given a higher cost to
represent that they are more difficult to perform.

None of these costs are very scientific but are mostly just
relative approximations for the amount of effort to act on a single
node. We can refine this later. One part that is not well represented
are literals. Literals are currently treated as having no cost, but
literals can have a cost when they are used in some contexts because we
might have to allocate a vector for the literal. Other times, literals
are kept as scalars such as when they are part of filtering operations.

For our current purposes, figuring out the different ways literals could
be represented didn't really seem to impact the plans very much so I
just left them as always free.

Another tweak to the way costs are written as part of the
`RelCostWriter`. This time, costs are moved to a different line from the
original relational node. This is to help differentiate between changes
to the cost and changes to the plan.

It might be the case that we make changes to how costs are calculated
that do not affect the plan. With the previous cost writer, it was
difficult to assess when the plan changed versus when the cost changed
by just looking at a diff. This was especially apparent for nodes that
were longer in length because of more parameters such as large numbers
of projections.

Now, costs are displayed on a newline above the node it is associated
with and prepended with a comment marker to differentiate them.

This should make it so we can identify the difference between costs
changing and plans changing.