[BSE-167] Correctly use the new condition expression instead of the old one (#5276)
PandasJoin wouldn't correctly utilize the passed in expression for the
condition when `copy` was used. This interfered with some planner rules
that utilize that method and it's just generally incorrect.

This fixes the copy method so the condition gets correctly replaced.
This fixes some of the volcano planner tests which were indirectly
impacted by the bug.