Increase the number of jobs used for Azure nightly (#5279)
Massively increases the number of jobs used for Azure nightly, and removes the parallel limit. This will be used in conjunction with the new scale set agents to decrease the amount of time that nightly takes.

I'm going to merge this after tonight's nightly run, just to be sure that there's no issues with the current setup with the scale set.