[BSE-270] Utilize join condition removal to filter for query 115 (#5339)
This utilizes the new planner rule that removes the condition from the
join and places it into a filter on top of the join. It also utilizes a
different version of the `FilterJoinRule.FilterIntoJoin` to work with
the new code from `PandasJoin`.

This can produce suboptimal plans because join doesn't have cost
computation. At the same time, it cannot produce invalid code which is
the underlying goal.

This also adds a small bugfix to the `JoinConditionToFilterRule`. It
turns out the logic we used for extracting the filter only works with
inner joins. It's possible to get it to work with various types of outer
joins, but that's not part of this change.