Keaton compilation debug utils (#5345)
This is a simple script I wrote to help me debug compilation errors in customer scripts. It essentially just takes a query, and returns a "type equivalent" query. This is useful for removing massive nested subquerries from customer code, when attempting to create a minimal reproducer.

Currently only works for simple select/with statements, which is all that's needed for right now.