[BSE-39] Support LIKE ANY/ALL with ESCAPE clause (#5309)
This adds support for LIKE ANY/ALL with the ESCAPE clause. This is done
by creating a customer operator for `LIKE ANY/ALL` that combines code
from both the like operator and the quantify operator. This custom code
ensures that code has the correct type and casts are performed if type
coercion is allowed. The parser utilizes these custom operators instead
of any of the operators in Calcite core.

After that, a convertlet table is used to transform the `SqlNode` to the
equivalent `RexNode`. This is used instead of modifying the
`SqlToRelConverter` in order to allow the `LIKE ANY/ALL` operator to be
rewritten as an expanded list of conditions joined by `AND/OR`.

Tests are added for the various stages that this affects. No changes to
code generation were made.

Not entirely thrilled with the error messages generated by invalid types
at the moment, but that will have to be a future improvement.

This also removes NOT LIKE ANY from the parser as it's not supported
snowflake syntax. This also removes it from testing since it's not
something that needs to be supported.