[BSE-229] Support additional escape sequences in the parser and code generator (#5361)
Backslash sequences in string literals are supported by Snowflake. This
updates the parser to include support for the backslash escape sequences
supported by Snowflake in order to support these queries as part of
LIKE. In addition to backslash, other escape sequences that are part of
string literals are now supported too.

This also updates the like code generation and the literal code
generation to use the `Expr` class instead of string builder with raw
code. This was done to allow for strings to be properly escaped and
rendered when escape sequences were present in the string.

This also adds query 33 and 55 to the list of tested queries and
fixes an issue with escape sequences in triple quoted strings such as
the rel node timer. query 33 is disabled temporarily as `LISTAGG`
is not supported by the code generator yet.