[BSE-335] Fix `ROW_NUMBER` distribution bug by using index bounds from input dataframe (#5369)
(As title)
The issue was that sort decides the bounds based on sampling, so during the shuffle, the ranks are not guaranteed to receive the same number of rows as the input dataframe (the total will be the same, but the distribution among the ranks may not be). To resolve this, we can simply keep track of the index boundaries (essentially cumulative sum of number of rows on each rank), and then use these boundaries during the "reverse" sort. This guarantees that the right amount of data is sent back to each rank and that it lines up perfectly with the input (since it's a simple RangeIndex). We just needed to add support for _bodo_chunk_bounds to df.sort_index, which was easy since we already support it for sort_values and the underlying infrastructure is the same for both.