[BSE-222] Add support for streaming Projections/Filters (#5376)
Updates BodoSQL to have a physical trait for BatchingProperty that is used to generate separate plans/code for streaming. This is initially blocked behind a module level flag so it won't be enabled on develop. However it can be enabled for testing.

This PR also include additional plan/codegen tests for the batchingProperty