[BSE-345] Support implicit casting in LEAST/GREATEST (#5387)
This supports implicit casting in LEAST/GREATEST.

This was previously supported because LEAST/GREATEST got rewritten into
case statements which supported implicit casting. When we changed these
to use their raw forms without using CASE to support filter pushdowns,
we also broke their ability to implicitly cast. In Calcite, the
`SameOperandTypeChecker` will ensure types are the same and will not
attempt to perform implicit casts. The `LEAST` and `GREATEST` functions
utilized this type checker and nothing else which meant that parameters
of different types would be caught as invalid.

In order to support this, we extend the operand type checker that was
previously used, check if type coercion is enabled, and then perform the
type coercion if it is possible. This follows a similar path to what
`CASE` would have done if it were used instead but with a function
instead.

This should case `CAST` expressions to be added correctly and should
allow `GREATEST` and `LEAST` to be used in places where an implicit cast
is allowed.