[BSE-306] Remove variable Shadowing in BodoSQL (#5409)
The primary goal of this PR is to remove all instances of variable shadowing in BodoSQL. IE:

```
df2 = df1.method(...)
df2 = df2.other_method(...)
```
In the process of doing this, I did a little bit of refactoring, in order to make it clearer which functions actually generating intermediate assignments. Generally, this involves passing less raw strings back and forth between functions. Most of the touched code still uses raw string expressions, I didn't want to balloon the scope of this PR, since this is a blocker for streaming codegen. There's still a huge amount of followup work to be done, but this should hopefully make the codebase a bit more readable.

I've made a notable exception for handling binops, case, and window functions, since all of these are under active changes, and I don't want to add any unnecessary conflicts to an already large PR.