Refactor IR expressions to be more generic and utilize Kotlin better (#5470)
This change modifies the IR values to utilize Kotlin a bit better and
act more generic. In particular, one new IR node, the `DelegateExpr`,
allows composing new IR's through existing IR primitives.

In general, I'd caution the use of `DelegateExpr`. It's likely better to
just create helper functions that utilize the IR primitives rather than
to add delegate expressions.

This also simplifies some of the IR arguments and improves some of the
usability. An example is that now `Method` is just an attribute access
passed to `Call` and `Call` is now capable of taking an `Expr` as the
callee. The `Dict` literal now takes in a list of pairs rather than
taking in two lists and some of the emit functions are now using Kotlin
helper functions in a more functional style than they were previously.

The `data class` has also been removed. It hasn't been useful.