[BSE-505] Fix various regressions in the volcano planner (#5527)
Fix various regressions in the volcano planner

This contains numerous fixes to enable the volcano planner to be used as
an equivalent of the existing heuristic planner.

RelMdSize has been updated to provide some suitable defaults for
`VARCHAR` and `VARBINARY`. It has also been given a suitable value for
`UNKNOWN` just to ensure it doesn't completely mess up the estimation of
sizes for types. This will help prevent strange numbers from appearing
when attempting to retrieve the size of rows or columns.

In order to simplify the construction of the `RelMetadataProvider`, a
single class `PandasRelMetadataProvider` now exists to encapsulate that
logic. The actual contents haven't changed but they are now not
duplicated in two locations.

The test harness has been fixed to always use this custom
`RelMetadataProvider` so the plan cost outputs are consistent with what
the planner was seeing. This was a visual error and didn't affect the
results. Still, I spent quite some time trying to understand why the
numbers weren't consistent with the output. The `RelMetadataProvider`
was being reverted back to the one used by the Heuristic Planner at some
point and the custom logic for `RelMdSize` was not being used in the
explain output.

The `RexCostEstimator` has been fixed to consider the evaluation of the
partition keys or order keys. Previously, these were accidentally
skipped so complex expressions in these positions were ignored and the
planner couldn't recognize that it would be more efficient to only
compute them in one location.