[BSE-559] Unit tests for edge cases related to handling of long strings in `ChunkedTableBuilder` (#5589)
(As title)
Tested indirectly by doing a simple join (essentially a vertical concat) and using inputs to force certain behavior.
There's a simple change in the BufferPool Cython interface which exposes the new function in C++ to get the size of the smallest SizeClass. This was required for testing purposes.