Modify PandasRel implementor and refactor PandasFilter to use it (#5498)
The PandasRel code generation has been fleshed out a bit so it can
actually be used. The `Implementor` acts as a way to expose behaviors
from the `PandasCodeGenVisitor` in a way that's more controlled
utilizing methods to visit inputs and then another method to trigger
the code generation context.

This should allow us to move some of the code generation in the `emit()`
functions for various transformations. Moving the code generation to
these classes should allow us to divide portions of the code generation
into different functions more easily which should make the core code
generation more understandable.

As a proof of concept, PandasFilter has been modified to utilize this
method and the filter code generation in other parts of the codebase
have been moved there.

This also fixes a bug with node caching for filter operations. Filter
operations were previously checking the cache, but they would never
insert nodes into the cache. This caused these to be guaranteed cache
misses. Filter operations now use the common code in `PandasRel` which
handles this for them.