[BSE-506] Utilize table size in the planner for snowflake tables (#5585)
This change introduces a method of retrieving the number of rows for a
specific table when the snowflake catalog is utilized. This will invoke
a `SELECT COUNT(*)` query on the given table to retrieve the number of
rows. Included in this is the ability for Calcite to memoize the result
so it isn't computed multiple times by the planner.

This is tested in the unit tests through the use of mocks and the
Java integration tests have configured a few tables to use different
default values.