[BSE-586] Utilize RelRoot when performing code generation (#5608)
This changes the way we handle the `RelNode` to utilize `RelRoot`
instead for code generation.

In Calcite, the planner rules will guarantee that equivalent nodes will
have the same indices and types, but it won't guarantee names. Instead,
Calcite expects the end-user to use `RelRoot` to determine the real
names of the output type.

For optimization, instead of using `project()` to create a projection
from the `RelRoot` at the beginning, we now pass the `RelNode` attached
to the `RelRoot` for optimization and keep replacing it with the new one
as we go through the chain. When we get to code generation, we now
invoke `pandasProject()` to transform the `RelRoot` to a `RelNode` where
the topmost node will be a `PandasProject` if required. This logic will
be skipped if it's a DML node (which don't follow these rules) or if the
projection is trivial for both the names and indices.

This make the code generation more reliable and allow us access to more
of the planner rules that don't necessarily keep the field names intact.