[BSE-498] Refactor visitAggOverOp to detangle it from the visitor and translator (#5535)
[BSE-498] Refactor visitAggOverOp

This refactors the `RexOver` handling so it is handled outside of the
`RexToPandasTranslator` and modifies the code generation to add a bit
more structure to it and use `Expr` nodes.

The biggest change is that `RexOver` nodes are now removes from
`RexNode` outside of the normal code generation. That means that filters
and projects, which are the only ones that support this node, will now
extract the window aggregates and perform the code generation to produce
the results before `RexToPandasTranslator` does anything.

This is effectively what was happening before but this involves less
connected parts between different sections of code. In the past, the
`RexToPandasTranslator` had to keep a bunch of state about what context
`RexOver` was found in and pass information about which columns were
referenced so it could eventually generate the apply function. Now, it
just does that ahead of time and the logic is removed from
`RexToPandasTranslator` and `PandasCodeGenVisitor`. `WindowAggregate`
will determine which columns are referenced, generate the appropriate
dataframe, generate the apply function, and invoke the call. It will
then retrieve each of the results and put it into a series that can
later be accessed.

The results are then replaced with `RexLocalRef` and
`RexToPandasTranslator` has been given support for local refs. These are
local variables that are primarily used by Calcite for code generation.
Local refs are not produced by any SQL node and are transformations just
to enable this kind of thing.

Fusing is also done in this area and fusing takes into account whether
the group is capable of being placed into a window function or if it
needs to generate the apply function.

In addition, code generation of individual aggregates has been moved
from a large function to individual functions for each windowed
operation. When they have something in common, they use a helper
function to produce the results. This should make it easier to find the
code generation for a specific function and easier to follow the code.

The primary benefit of this is we no longer need the mix in logic from
`RexToPandasTranslator` and `PandasCodeGenVisitor`.
`RexToPandasTranslator` still references the latter, but it uses it for
information rather than code generation. That means we can use
`RexToPandasTranslator` in a more generic way rather than only at the
topmost level of code generation. It also opens up the possibility of
using the scalar context on its own rather than as part of the column
context.

There's still some continued work to do in this area. The code is still
not as clear as I'd like it to be and some of the functions are given
fairly ambiguous names. At the same time, this should provide a basis
for more gradual improvements in the future for code readability.