[BSE-574] Support BodoSQL codegen for streaming groupby (#5604)
* initial wrappers and plumbing for streaming groupby

* minor refactor

* module setup

* fix minor typo

* fix out table type

* python wrapper fixes

* typing transform for groupby state

* fix dict builders

* implement basic groupby sum

* initial parallel support

* add shuffle code

* Support parallelism in Python

* pass function type arguments

* pass function info to C++

* refactor groupby key init function

* reuse existing update infrastructure for input batch

* call proper combine function

* add start_row to all kernel cases

* add parallel combine support

* [BSE-574] Support BodoSQL codegen for streaming groupby

* Modify PandasRel implementor and refactor PandasFilter to use it (#5498)

The PandasRel code generation has been fleshed out a bit so it can
actually be used. The `Implementor` acts as a way to expose behaviors
from the `PandasCodeGenVisitor` in a way that's more controlled
utilizing methods to visit inputs and then another method to trigger
the code generation context.

This should allow us to move some of the code generation in the `emit()`
functions for various transformations. Moving the code generation to
these classes should allow us to divide portions of the code generation
into different functions more easily which should make the core code
generation more understandable.

As a proof of concept, PandasFilter has been modified to utilize this
method and the filter code generation in other parts of the codebase
have been moved there.

This also fixes a bug with node caching for filter operations. Filter
operations were previously checking the cache, but they would never
insert nodes into the cache. This caused these to be guaranteed cache
misses. Filter operations now use the common code in `PandasRel` which
handles this for them.

* [BSE-541] Handle Snowflake Gateway Timeouts in Flaky (#5586)

* [BSE-500] Fix Mac x86 Builds (#5584)

* [BSE-478] Support dict encoding in nested loop join  (#5581)

(As title)

* [BSE-506] Utilize table size in the planner for snowflake tables (#5585)

This change introduces a method of retrieving the number of rows for a
specific table when the snowflake catalog is utilized. This will invoke
a `SELECT COUNT(*)` query on the given table to retrieve the number of
rows. Included in this is the ability for Calcite to memoize the result
so it isn't computed multiple times by the planner.

This is tested in the unit tests through the use of mocks and the
Java integration tests have configured a few tables to use different
default values.

* [BSE-503] Fix Mac M1 Builds (#5596)

* [BSE-317] c++ changes for nullable timestamps (#5463)


---------

Co-authored-by: pintaoz2 <pintao@bodo.ai>

* add unit test

* fix test

* append key rows to running buffers properly

* refactor reserve functions

* pass proper input columns to ColSet

* refactor get_groupby_output_dtype() to avoid side effects

* allocate and access update/combine columns properly

* add comments [run ci]

* create function for streaming aggregation

* add comments

* more comments

* minor comment

* add combine offsets to all colset classes [run ci]

* minor fixes

* [run ci]

* update PandasCodeGenVisitor.java

* get init state input variables

* get init state input variables

* update comments

* add visit input

* add flag to enable streaming groupby

* update PandasCodeGenVisitor

* update codegen output

* set the flag to true and [run CI]

* [run CI]

* [run CI]

* [run CI]

* fix AppendRow calls [run ci]

* fix typo

* fix typo

* improve byte aligned bitmask init [run ci]

* fix linux error [run ci]

* [run CI]

* [run CI]

* address comment

* address comment

* [run CI]

* address comments

* [run CI]

* [run CI] set enable_groupby_streaming to False

---------

Co-authored-by: Ehsan Totoni <ehsan.tn@gmail.com>
Co-authored-by: Jonathan A. Sternberg <jonathan@bodo.ai>
Co-authored-by: Srinivas Lade <srinivas@bodo.ai>
Co-authored-by: Isaac Warren <42949629+IsaacWarren@users.noreply.github.com>
Co-authored-by: tathougies-bodo <127988135+tathougies-bodo@users.noreply.github.com>