[BS3-53] Support mode as a groupby aggregation (#5676)
Adds support for the aggregation function `MODE` as a groupby aggregation. Window support is already handled, and no-groupby (+ `HAVING`) support will be in a seperate PR. 

This function needs to support almost every dtype and array type, so a great deal of templating and dtype/arrtype agnostic operations are required. 

Nullable and Numpy arrays go to the same dtype-blind implementation that builds a hashtable for each group mapping each unique element to its count, and then picks the one with the highest count.

String arrays do the same but by constructing a `std::string_view` for each value in the array and after the mode is found, placing the actual string in the result array.

Dictionary encoded arrays simply find the mode amongst the index array, then return an identical dictionary encoded to the input array but with the indices swapped out for the index of the mode within each group. Potential followup: remove indices from the final array if they are not used?