[BSE-666] Customize row counts for the TPCH schema to the real values (#5742)
This allows the schema to be customized with row counts that would be
typical in a real planning environment. The DSL has been updated to
include some of these as attributes.

The cost output has also been changed a bit to accommodate this change.
It now emits scientific notation in multiples of 3 (aka thousands,
millions, billions, etc) and does not emit scientific notation at all if
the exponent is zero. It will also not emit scientific notation with an
exponent that is not a multiple of 3 to keep reading the costs easier.

The cost code has been refactored a bit to remove the duplication that's
present with the decimal formatter in both cost and the `RelCostWriter`
because the logic is now a bit more complex.

In addition to the above, the unoptimized plans have now been changed to
emit the output after the first transformation program has run. This was
originally just subqueries but will likely be extended to more in the
future as a preprocessor program.

Originally, this included the raw output of the `SqlToRelConverter` to
allow seeing the 1:1 transformation from there. This has been
increasingly less helpful while it's become increasingly more important
to see how the optimizer changes the plans after subqueries are removed.

Plans have been updated accordingly.