[BSE-736] Fixes issue in streaming groupby codegen (#5762)
he date tests listed in the issue are fixed by the changes, but the fact that this error only ocured for tests specific to dates was a complete red herring. I expect this change should fix a number of the other existing bugs. I've set the flags and am running nightly HERE