[BSE-608] Support count functions with groupby.window (#5697)
Adds support for the functions COUNT, COUNT(*) and COUNT_IF in the groupby.window framework. Entails the following:

    Porting all existing groupby.window logic and tests to dedicated files
    Adding new templated algorithms for window functions that do partition-wide aggregation, prefix frames, suffix frames, or sliding frames.
    Adding a new class WindowAggfunc with templated helper methods for the algorithms to use, with implementaitons created for the 3 newly supported ftypes.

The 5 helper functions are templated on the window function ftype, array type, input typename, input dtype, and an enum of the type of frame (sliding vs cumulative vs no_frame). There are 5 methods to implement:

    init: initalizes/resets the data within to begin the accumulation.
    enter: adds a value to the accumulator.
    exit: remove a value from the accumulator (only required for implementations that support sliding frames).
    compute_partition: using the values in the accumulator, fills the entire partition with the single value resulting from the partition-wide aggregation.
    compute_frame: using the values in the accumulator, computes the value for the aggregation at one moment in time and places it in the corresponding location of the output array. (only required for implementations that support sliding or cumulative frames)

The reason for this setup is to provide flexibility for functions whose behavior may alter with dtype / array type, and for functions where a significantly faster implementation is possible for partition-wide / cumulative frames but not for sliding ones (e.g. min/mode). Any function that does not need to support window frames but does need the partition-wide aggregation can do so by only implementing init, enter and compute_partition; the C++ code will fall back on the default implementations for the other 2 methods (which just raise errors).