[BSE-766] Simple single-core regression test for streaming hash join (for milestone 2 PRs in particular) (#5790)
(As title)
Test for checking regression in streaming hash join. This is meant to detect serious regressions in single-core performance, but can be used for the multi-core case as well. This test can be updated in the future as the the BodoSQL code-gen starts generating more efficient code in general.
It's recommended to run this on a single node on the platform (e.g. c5n.18xlarge) to get consistent results.
It's a single hash join on two tables from the TPCH dataset.