[BSE-666] Utilize multi join for organizing join orderings (#5748)
This utilizes MultiJoin rules to group joins together and then determine
an optimal ordering. In order to facilitate this, custom logical nodes
have been created in order to propagate our cost changes to the logical
nodes since the MultiJoin rule only operates on logical nodes.

There are some caveats to this implementation and it's likely not an
ideal implementation.

First, the join ordering assumes the left side is the streaming side
while the right side is the batch/in-memory side. This is the opposite
of how our code handles these, but just swapping these in code
generation could produce a problem in the output so the
`JoinCommuteRule` is used to fix it after the fact. Still, if this is
true, this can result in an unoptimal path because `LoptOptimizeJoinRule`
can make an irreversable decision based on the wrong ordering.

Second, the ordering it chooses is unclear to me for why it chooses
what it does. The algorithm choices are not documented clearly so I
don't know why it chooses the ordering it does. This will require
further investigation as we should be able to answer questions about how
the ordering is determined.

Third, projections that eliminate columns are not currently considered
at the time that the join ordering is chosen. That means it's possible
for the ordering to treat a table with a large number of columns that
are never selected as being a higher cost than a table with a small
number of columns but all of them are selected. There are further
enhancements in Calcite that allow projections to be pushed into a
`MultiJoin` that may be worth exploring. Alternatively, we can create
our own or find some way to ensure this information is present.

We will want to revisit other ways to perform joins but this should at
least improve the join ordering a bit.