[BSE-648] Port any_value to groupby.window (#5740)
Adds support for the function ANY_VALUE in the groupby.window framework. Entails the following:

    Repeating procedures developed in 

    [BSE-608] Support count functions with groupby.window #5697 for the new ftype
    Adding support in the window frame framework for having string & dictionary encoded outputs
        Adding a new function cleanup that is run after all partitions are finished. This funciton is responsible for taking any intermediary representations and converting them to the final string/dict array.
        Expanding the agnostic utilities in _array_utils.h to include string/dict arrays.
    Adjusting codegen tests for ANY_VALUE (and the rest of the value functions) to reflect the full set of allowed behavior for each function.

The reason for this setup is to provide flexibility for functions whose behavior may alter with dtype / array type, and for functions where a significantly faster implementation is possible for partition-wide / cumulative frames but not for sliding ones (e.g. min/mode). Any function that does not need to support window frames but does need the partition-wide aggregation can do so by only implementing init, enter and compute_partition; the C++ code will fall back on the default implementations for the other 2 methods (which just raise errors).