Remove the calcite fork (#5759)
This copies the files that were modified in the fork into our own
project. It utilizes the shade plugin to overwrite the class files from
the upstream dependency which is effectively the same as maintaining the
fork but without jitpack and only the files that have been modified are
present.

This copies over the files as-is that are different. Some of the changes
are trivial and can be removed.

This also includes the license and the attribution in our shipped
version of Bodo to comply with the license. We will still need to add
the "this was modified" header to any of the modified files. The primary
thing needed for the Calcite license is including a textual distribution
of the `NOTICE` file from the upstream Calcite repository.

The assembly plugin is removed since it overlaps with the shade plugin
for creating an uberjar but the shade plugin allows us to resolve class
conflicts for ones we overwrite. We can populate that list by using the
following command:

    $ mvn clean compile
    $ cd bodosql-calcite-application/target/classes
    $ find org -type f | sort | sed 's/\(.*\)/<exclude>\1<\/exclude>/g' | pbcopy

Note that the `org` is intentional instead of `org/` which would be put
there by shell completion. It prevents an extra slash from being
included. `pbcopy` is also a mac only command to copy something to the
clipboard. This list is then pasted into the `pom.xml` from the editor
with the correct indentation.

The main class was also removed from the `pom.xml` file. It's possible
to set a main class for the uberjar from shade, but this main class
wasn't used anyway as py4j uses an internal main class automatically.
Just removing that file as its existence was confusing and misleading.