Fix maven incremental compilation (#5828)
Maven incremental compilation had a few aspects to it that were broken
and ignored.

Kotlin incremental compilation was disabled. This was because Kotlin
incremental compilation commonly gets broken and the Gradle support for
Kotlin is much better. At the same time, upgrading Kotlin to the latest
version and enabling it seems to result in it working.

Java had the same build step defined twice accidentally and was running
into an error with `package-info.java`. Since that file exists solely to
ensure the package exists for the parser to be generated, it never
creates a class file which causes Java to think the build is constantly
stale. The answer to fix that is here: https://stackoverflow.com/a/57004351

The fmpp build plugin depended on a drill-root which had an outdated
reference to conjars.org. conjars.org was taken down by the owner in
2021 (https://groups.google.com/g/cascading-user/c/qAQvajapawI?pli=1).
Upgrading the plugin seems to have resolved the issue.