First iteration for removal of fork modifications (#5829)
Trivial modifications have been removed. First, removed any files with
trivial modifications (i.e. formatting) that were probably accidentally
done by an editor. Second, there were some changes that could be removed
if a smaller change was made in another modified file. In particular,
`SqlSelect` had a parameter added to it. If an overloaded constructor
was included, we would not have to modify callsites that did not use the
new parameter.

The prepare package has been removed. We now use `BodoCatalogReader`
which has access to the protected member that we were attempting to
expose and we no longer use `PlannerImpl`. We also do not use
`FrameworkConfig` for configuring the catalog reader (although we do use
it for other stuff).

Changes to the `SqlStdOperatorTable` have been ported to
`SqlBodoOperatorTable` and put higher in the precedence order so those
overloads are used instead. `SqlBodoOperatorTable` also uses the
reflective operator table mechanism that allows `SqlStdOperatorTable` to
self-discover new functions. That means new functions can be added by
just creating a new member variable.

Removes changes to `OperandTypes` and `ReturnTypes`. These changes were
capable of being kept locally and attached to the correct operator in
our own operator table. At the same time, they utilized constructors
that were package-private so they are still in the apache package
namespace, but no longer require an override.

The random function does not require customizing `SqlOperator`. The
random function should have just been a custom operator type that
overrode the `isDeterministic()` function. It was incorrect to modify
`SqlOperator`.

Moves a change in RelOptUtil to its own method and utilizes that instead
within `SqlToRelConverter`. This doesn't remove the change as would be
done in a more ideal version of this, but it allows us to invoke the
copied method from a new class instead of modifying the one in upstream.
We modify `SqlToRelConverter` to call that new one instead.

The `REGEXP` operator has been removed in favor of utilizing `RLIKE`.
This operator is the equivalent of `RLIKE` and doesn't have a function
version so there's no need to have an operator. The parser has been
changed to turn `REGEXP` into `RLIKE` and this leads to the code
generation being more simple and the removal of `SqlKind.REGEXP` and any
changes to `SqlLikeOperator` or `SqlLibraryOperators`.