Make RexNamedParam an SqlOperator (#5834)
This converts named parameters to be an operator and modifies the parser
to substitute the call for any reference to named parameters.

This also includes some custom logic for validation with offset and
fetch. Traditionally, offset and fetch don't allow calls and only allow
literals or dynamic parameters. This adds an additional validation step
that handles calls in this area.