Remove INTERVAL_WEEK and add support for QUARTER (#5838)
The week interval can be represented as 7 days. The area of the interval
parser that deals with this now treats it that way and otherwise treats
it as a day interval.

There was no area where we treated week in a special way from a typing
perspective.

The upstream version of Calcite has support for WEEK and QUARTER, but
the functionality is broken. This change includes the changes required
to fix it that have been sent upstream here:
https://github.com/apache/calcite/pull/3329