[BSE-778] Support optional extra arguments for CEIL and FLOOR (#5815)
Adding support for FLOOR and CEIL with an optional precision argument. Requires messing with the operators used because the Floor/Ceil operators in the standard operator table will not work with 2 arguments without making several Calcite changes. This is possible to do, but it's much simpler to just add a new floor/ceil operator in our own operator tables.