Removed a rename test (#5855)
I'm removing a test that fails in the streaming group by case. Essentially, we don't generate a call to df.rename in the streaming case, because we pass the aggregation input && get the output as tables. Therefore, when we eventually package this back into a Dataframe, we pass the new names directly to init_dataframe, which avoids the call to dataframe.rename that is being checked for in this test.

I can move this to a maven test if needed, but this seems like a pretty pointless test that I'd rather just remove entirely, especially because I expect the upcoming move to using tables as the default intermediate value instead of dataframes will break this again.